modularized mess in main into seperate files
